Add `make dev-arti` as an alternative to C-Tor
Arti is Tor's RIIR effort, and while it is 1.x, the onion service
support is still experimental.

Given that, it is to our (and Tor's) benefit for us to try using Arti in
development modes to provide feedback. To that end, using Arti is now
possible, in parallel, to the existing C-Tor support.

Running `make dev-arti` should give exactly the same experience as the
existing `make dev-tor` endpoint, aside from the inital launch, which
will be slower since it needs to compile Arti from scratch.

As an intentional choice, we don't version pin Arti since this is an
experimental thing, so it's more beneficial we just automatically stay
on the latest version and fix breakage instead of needing to constantly
bump versions.

Note that `make dev-tor` is entirely independent from this, so the two
commands will serve the dev server on different onion addresses.

As far as the Arti code itself, generating a config.toml file seems much
cleaner than adding to `torrc` and generating `.auth` files.