Run securedrop-remove-packages hourly instead of daily
This is an incredibly quick thing so let's run it more frequently so it
is taken care of sooner and reduces the risk of the check script
complaining that ufw is still present.

https://www.freedesktop.org/software/systemd/man/latest/systemd.time.html#Calendar%20Events
indicates that "hourly" is a supported shorthand.

(cherry picked from commit 0a163ef06ac3341a80fc2f8650573878aee2214e)