Make noble-migration playbook smarter for SSH-over-Tor handling
Previously, the ansible playbook didn't actually understand the
sequencing - it was just a dumb, hardcoded sequence of two reboots.

We recently learned that with SSH-over-Tor there's a race condition, you
might end up with 3 disconnections instead of 2 because you'll lose SSH
when Tor is upgraded. In some cases if things are slow enough, you might
only reconnect after the second reboot. So we need to be able to
dynamically handle 2 or 3 disconnections.

The playbook is now aware of the most recently completed state, as well
as the position in the overall run.

In human-readable pseudocode, the logic roughly is:

* Get current finished state
* If finished_state == None: (i.e. not started)
  * Kick off the migration.
  * Wait for a reboot (should switch to PendingUpdates here)
  * Get current finished state
* If Reboot > finished_state >= PendingUpdates: ("Phase 2")
* Wait for a reboot (should switch to Reboot here, unless SSH-over-Tor
  disconnects first)
  * Get current finished state
* If SSH-over-Tor and Reboot > finished_state >= PendingUpdates: ("Phase 2.5")
  * Wait for a reboot (should switch to Reboot here)
  * Get current finished state
* If Done > finished_state >= Reboot: ("Phase 3")
  * Wait for finished state to be Done
  * Set current finished state to Done
* If finished_state != Done:
  * Fail :(

Because the logic is fully aware of the migration state, if you e.g.
start the playbook, the upgrade begins, you Ctrl+C, and then restart the
playbook, it'll just work because it'll drop you into the correct
phase's logic instead of starting from the beginning.

Fixes #7483.

(cherry picked from commit 07673333f15a86f8f3bec021801852b1d5815e66)